examples/jax_styletransfer/requirements.txt
optax
dm-haiku
Pillow
h5py